pyupgrade and isort
