--- MANIFEST.in
@@ -1,3 +1,0 @@
-include tests
-include requirements.txt
-recursive-include bgmi *
