+++ bgmi/front/resources.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 import datetime
 import os
 from collections import defaultdict
