+++ bgmi/lib/fetch.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 from bgmi.config import DATA_SOURCE
 from bgmi.utils import print_error
 from bgmi.website import bangumi_moe, mikan, share_dmhy
