+++ bgmi/lib/models.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 import os
 import time
 from collections import defaultdict
