+++ pyproject.toml
@@ -38,12 +38,11 @@ bgmi_http = 'bgmi.front.server:main'
 [tool.poetry.dependencies]
 python = "^3.5"
 
-six = '>=1.10.0'
-beautifulsoup4 = '>=4.4.0'
-requests = '>=2.9.1,<3.0.0'
-tornado = '>=4.0.2'
-icalendar = '>=3.11.3'
-peewee = '>=3.3.4'
+beautifulsoup4 = '^4.4.0'
+requests = '^2.9.1'
+tornado = '^5.0.2'
+icalendar = '^3.11.3'
+peewee = '^3.3.4'
 
 [tool.poetry.dev-dependencies]
 pre-commit = { version = '2.2.0', python='>=3.6.1' }
