--- requirements.txt
@@ -1,6 +1,0 @@
-six>=1.10.0
-beautifulsoup4>=4.4.0,<5.0.0
-requests>=2.9.1,<3.0.0
-tornado>=6.0.0,<=7.0.0
-icalendar>=3.11.3,<4.0.0
-peewee>=3.3.4,<4.0.0
