+++ tests/test_http_api.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 import json
 import logging
 import os
