+++ tests/test_script.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 import requests
 from bgmi.script import ScriptBase
