+++ tests/test_utils.py
@@ -1,5 +1,3 @@
-from __future__ import print_function, unicode_literals
-
 import unittest
 
 from bgmi.utils import parse_episode
